refactored listings page as container
